+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version 0.0.8 (master)
 
+ - Added Adafruit NeoPixel support
  - Support for up to 4 LED output pins (TODO add option to configure strip segments in the UI)
  - Added LED Matrix and Strip support to ESP32 (wled_esp32_controller, compatible with many WLED LED controllers or my own version [EasyEDA schematics and PCB](https://oshwlab.com/sascha23095123423/wled-controller_copy))
  - Fixed a crash in RTCMemoryManager for ESP32